Adding the jobs apps
I've added the jobs apps to the djangogirls website project.
I've included the jobs urls.py in the main urls.py file adding it before the events as otherwise "jobs" in URL
was treated as a name of a city and was redirected to the events.
I've also created a simple view and template so one can see that everything is working.