Edited some phrasing, changed link
I feel like the link to donate should be in the call to action, so I added it there, and removed the other link. Feel free to reject if you disagree. The reason I did this is that when I saw this on the page, I totally missed the next sentence. My brain expected the "make a donation" text to be a link, and when it wasn't one, I stopped reading.