Create event from event application
* I extracted logic that deploys new city to `create_event` module, as
it is complecated and I think it makes sense to have it in separate
file. Thanks to that it is also easier to test.
* add function that creates `Event` based on `EventApplication`
* I moved some bits that copy content and menu from default event to
newly created one from `core.forms` to `create_event.py` - I don't see
any reason why we need a form here at all. We might be able to clean up
the form later on if it's not used anywhere else

Ticket #342